requirements.txt
colorama==0.4.6
instagrapi==1.16.30
python-dotenv==0.21.0
